Move project creation code from SystemManager to ProjectCreator
Bring all project creation code into one place. Turn ProjecCreator into
a fluid api class.
